Moved C syscall handlers to `host/syscall/handler/`
