+++ src/main/host/syscall/handler/file.c
@@ -3,7 +3,7 @@
  * See LICENSE for licensing information
  */
 
-#include "main/host/syscall/file.h"
+#include "main/host/syscall/handler/file.h"
 
 #include <errno.h>
 #include <fcntl.h>
